Split FacetTest selenium class in three. This makes it easier to read through the run logs to find where errors have occurred and possibly reduces chances of tests failing randomly because the browser overheats or whatever it is that causes random test failures.
